+++ App.js
@@ -1,3 +1,4 @@
+import 'react-native-gesture-handler'; // должен быть в самом верху
 import { StatusBar } from 'expo-status-bar';
 import {
   ImageBackground,
@@ -7,56 +8,57 @@ import {
   TouchableWithoutFeedback,
   View,
 } from 'react-native';
-import { RegistrationScreen } from './Screens/RegistrationScreen';
+import { RegistrationScreen } from './Screens/auth/RegistrationScreen.js';
 import { useFonts } from 'expo-font';
-import { LoginScreen } from './Screens/LoginScreen';
+import { LoginScreen } from './Screens/auth/LoginScreen.js';
 
-import imageBG from './Screens/img/photo-bg.jpg';
+import imageBG from './assets/img/photo-bg.jpg';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
+
+const AuthStack = createStackNavigator(); // для групировки экранов и рендера отдельных экранов
 
 export default function App() {
   const [fontsLoaded] = useFonts({
-    'Roboto-700': require('./Screens/fonts/Roboto-Bold.ttf'),
-    'Roboto-400': require('./Screens/fonts/Roboto-Regular.ttf'),
-    'Roboto-500': require('./Screens/fonts/Roboto-Medium.ttf'),
+    'Roboto-700': require('./assets/fonts/Roboto-Bold.ttf'),
+    'Roboto-400': require('./assets/fonts/Roboto-Regular.ttf'),
+    'Roboto-500': require('./assets/fonts/Roboto-Medium.ttf'),
   });
+
   if (!fontsLoaded) return null; // сначала загрузка шрифтов, потом все остальное
 
   return (
-    <View style={styles.container}>
-      <TouchableWithoutFeedback
-        onPress={() => {
-          console.log('Hi'), Keyboard.dismiss();
-        }}
-      >
-        <ImageBackground style={styles.imageBg} source={imageBG}>
-          <KeyboardAvoidingView
-            style={styles.wrapperForm}
-            behavior={Platform.OS == 'ios' ? 'padding' : 'height'}
-          >
-            <RegistrationScreen />
-            {/* <LoginScreen /> */}
-          </KeyboardAvoidingView>
-        </ImageBackground>
-      </TouchableWithoutFeedback>
-      <StatusBar style="auto" />
-    </View>
+    <NavigationContainer style={styles.container}>
+      <AuthStack.Navigator initialRouteName="LoginScreen">
+        <AuthStack.Screen
+          name="Registration"
+          component={RegistrationScreen}
+          options={{ headerShown: false }} // скрывает header
+        />
+        <AuthStack.Screen
+          name="Login"
+          component={LoginScreen}
+          options={{ headerShown: false }}
+        />
+      </AuthStack.Navigator>
+    </NavigationContainer>
+
+    // <>
+    //   <View style={styles.container}>
+    //     <RegistrationScreen />
+    //     {/* <LoginScreen /> */}
+
+    //     <StatusBar style="auto" />
+    //   </View>
+    // </>
   );
 }
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
+    backgroundColor: '#ffeafeff',
     // alignItems: "center",
     // justifyContent: "center",
-  },
-  imageBg: {
-    flex: 1,
-    resizeMode: 'cover',
-    justifyContent: 'flex-end',
-  },
-  wrapperForm: {
-    borderTopLeftRadius: 25,
-    borderTopRightRadius: 25,
-    backgroundColor: '#fff',
   },
 });
